+++ test/dummy/config/application.rb
@@ -17,6 +17,8 @@ require "action_cable/engine"
 # Require the gems listed in Gemfile, including any gems
 # you've limited to :test, :development, or :production.
 Bundler.require(*Rails.groups)
+require "recording_studio_publishable"
+require "recording_studio_attachable"
 
 module Dummy
   class Application < Rails::Application