+++ test/dummy/config/initializers/recording_studio_admin.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+RecordingStudioAdmin.configure do |config|
+  config.default_mount_path = "/admin"
+  config.authentication_method = :authenticate_user!
+  config.current_actor_method = :current_user
+  config.access_recording_resolver = lambda { |_context| nil }
+  config.site_admin_recording_resolver = config.access_recording_resolver
+end