+++ test/dummy/config/initializers/recording_studio_attachable.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+RecordingStudioAttachable.configure do |config|
+  config.max_file_size = 25.megabytes
+end